+++ src/standx_sdk/domain/__init__.py
@@ -1,6 +1,14 @@
 from .account import AccountApi
+from .account_views import PositionsApi, TradesApi
 from .markets import MarketsApi
 from .order_recovery import OrderStateReconciler
 from .orders import OrdersApi
 
-__all__ = ["AccountApi", "MarketsApi", "OrderStateReconciler", "OrdersApi"]
+__all__ = [
+    "AccountApi",
+    "MarketsApi",
+    "OrderStateReconciler",
+    "OrdersApi",
+    "PositionsApi",
+    "TradesApi",
+]